ORV2-2623 - FE: Permission Matrix in Applications In Progress tab (#2032)
