+++ frontend/src/features/permits/components/dashboard/ApplicationStepPage.tsx
@@ -30,6 +30,7 @@ import {
   ApplicationStepContext,
   ERROR_ROUTES,
 } from "../../../../routes/constants";
+import { usePermissionMatrix } from "../../../../common/authentication/PermissionMatrix";
 
 const displayHeaderText = (stepKey: ApplicationStep) => {
   switch (stepKey) {
@@ -65,7 +66,7 @@ export const ApplicationStepPage = ({
   const enableMFP = featureFlags?.["MFP"] === "ENABLED";
   const enableSTFR = featureFlags?.["STFR"] === "ENABLED";
   const enableQRFR = featureFlags?.["QRFR"] === "ENABLED";
-  
+
   // Query for the application data whenever this page is rendered
   const {
     applicationData,
@@ -104,7 +105,9 @@ export const ApplicationStepPage = ({
   // Currently onRouteBC only handles TROS and TROW permits
   // other permit types are only allowed if its feature flag is enabled
   const isPermitTypeAllowed = () => {
-    const allowedPermitTypes: string[] = ([PERMIT_TYPES.TROS, PERMIT_TYPES.TROW] as string[])
+    const allowedPermitTypes: string[] = (
+      [PERMIT_TYPES.TROS, PERMIT_TYPES.TROW] as string[]
+    )
       .concat(enableSTOS ? [PERMIT_TYPES.STOS] : [])
       .concat(enableMFP ? [PERMIT_TYPES.MFP] : [])
       .concat(enableSTFR ? [PERMIT_TYPES.STFR] : [])
@@ -124,6 +127,14 @@ export const ApplicationStepPage = ({
     );
   };
 
+  const canEditIndividualApplicationInProgressDetails = usePermissionMatrix({
+    permissionMatrixKeys: {
+      permissionMatrixFeatureKey: "MANAGE_PERMITS",
+      permissionMatrixFunctionKey:
+        "EDIT_INDIVIDUAL_APPLICATION_IN_PROGRESS_DETAILS",
+    },
+  });
+
   const renderApplicationStep = () => {
     if (applicationStep === APPLICATION_STEPS.REVIEW) {
       return (
@@ -140,10 +151,12 @@ export const ApplicationStepPage = ({
   };
 
   if (
+    // add permissions matrix check here
     isInvalidApplication ||
     !isValidApplicationStatus() ||
     !companyId ||
-    !isPermitTypeAllowed()
+    !isPermitTypeAllowed() ||
+    !canEditIndividualApplicationInProgressDetails
   ) {
     console.error("The application cannot be displayed or edited");
     return <Navigate to={ERROR_ROUTES.UNEXPECTED} />;
