+++ frontend/src/features/permits/components/dashboard/PermitLists.tsx
@@ -12,15 +12,10 @@ import { applyWhenNotNullable } from "../../../../common/helpers/util";
 import { getCompanyIdFromSession } from "../../../../common/apiManager/httpRequestHandler";
 import { PERMIT_TABS } from "../../types/PermitTabs";
 import { useLocation } from "react-router-dom";
-
-interface PermitDashboardTab {
-  label: string;
-  component: JSX.Element;
-  componentKey: string;
-}
+import { TabComponentProps } from "../../../../common/components/tabs/types/TabComponentProps";
 
 export const PermitLists = React.memo(() => {
-  const tabs: PermitDashboardTab[] = [];
+  const tabs: TabComponentProps[] = [];
 
   const companyId: number = applyWhenNotNullable(
     (id) => Number(id),
@@ -28,14 +23,14 @@ export const PermitLists = React.memo(() => {
     0,
   );
 
-  const showApplicationsInProgressTab = usePermissionMatrix({
+  const canViewListOfApplicationsInProgress = usePermissionMatrix({
     permissionMatrixKeys: {
       permissionMatrixFeatureKey: "MANAGE_PERMITS",
       permissionMatrixFunctionKey: "VIEW_LIST_OF_APPLICATIONS_IN_PROGRESS",
     },
   });
 
-  if (showApplicationsInProgressTab) {
+  if (canViewListOfApplicationsInProgress) {
     tabs.push({
       label: "Applications in Progress",
       component: <ApplicationsInProgressList companyId={companyId} />,
@@ -43,14 +38,14 @@ export const PermitLists = React.memo(() => {
     });
   }
 
-  const showApplicationsInReviewTab = usePermissionMatrix({
+  const canViewListOfApplicationsInReview = usePermissionMatrix({
     permissionMatrixKeys: {
       permissionMatrixFeatureKey: "MANAGE_PERMITS",
       permissionMatrixFunctionKey: "VIEW_LIST_OF_APPLICATIONS_IN_REVIEW",
     },
   });
 
-  if (showApplicationsInReviewTab) {
+  if (canViewListOfApplicationsInReview) {
     tabs.push({
       label: "Applications in Review",
       component: <ApplicationsInReviewList />,
