+++ frontend/src/features/permits/helpers/mappers.ts
@@ -2,16 +2,6 @@ import { PermitsActionResponse } from "../types/permit";
 import { Nullable, Optional } from "../../../common/types/common";
 import { getDefaultRequiredVal } from "../../../common/helpers/util";
 import {
-  PERMIT_APPLICATION_ORIGINS,
-  PermitApplicationOrigin,
-} from "../types/PermitApplicationOrigin";
-
-import {
-  IDIR_USER_ROLE,
-  UserRoleType,
-} from "../../../common/authentication/types";
-
-import {
   Vehicle,
   VehicleType,
   VEHICLE_TYPES,
@@ -56,7 +46,7 @@ export const getSubtypeNameByCode = (
   if (vehicleType === VEHICLE_TYPES.TRAILER) {
     return trailerSubtypeNamesMap.get(typeCode);
   }
-  
+
   return powerUnitSubtypeNamesMap.get(typeCode);
 };
 
@@ -78,24 +68,4 @@ export const removeEmptyIdsFromPermitsActionResponse = (
     success: successIds,
     failure: failedIds,
   };
-};
-
-/**
- * Determine whether or not a given user can access/delete an application.
- * @param permitApplicationOrigin Permit application origin
- * @param role Role of the logged in user
- * @returns Whether or not the user can access/delete the application.
- */
-export const canUserAccessApplication = (
-  permitApplicationOrigin?: Nullable<PermitApplicationOrigin>,
-  role?: Nullable<UserRoleType>,
-) => {
-  if (!role) return false;
-
-  // CV/PA can only access/delete applications whose origins are not "PPC"
-  // Staff can access/delete any application they have access to (including each others')
-  return (
-    permitApplicationOrigin !== PERMIT_APPLICATION_ORIGINS.PPC ||
-    (Object.values(IDIR_USER_ROLE) as UserRoleType[]).includes(role)
-  );
 };
